record-api/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
python-jose==3.3.0
redis==5.0.1
python-dotenv==1.0.1
mysqlclient==2.2.4
python-multipart==0.0.9 